Allow an archive to be defragmented to save wasted space

+++ src/hed.rs
@@ -38,12 +38,9 @@ impl HED {
         P: AsRef<Path>,
     {
         // Create file if it doesn't exist yet
-        let is_file_new = if !path.as_ref().exists() {
+        if !path.as_ref().exists() {
             std::fs::File::create(&path).map_err(|_| ArchiveError::HEDReadError)?;
-            true
-        } else {
-            false
-        };
+        }
 
         // Read existing file
         let data = std::fs::read(&path).map_err(|_| ArchiveError::HEDReadError)?;
@@ -52,7 +49,7 @@ impl HED {
         let mut entries = Vec::<HEDEntry>::with_capacity((data_len / 12) as usize);
 
         // There will be nothing to deserialize if the file was just made
-        if !is_file_new {
+        if data_len != 0 {
             let mut cursor = Cursor::new(data);
             loop {
                 let entry = HED::try_deserialize_entry(&mut cursor)
